cognivoice-backend/requirements.txt
# --- Web Application & API ---
Flask==2.3.3
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
Flask-Bcrypt==1.0.1
Flask-Cors==4.0.0
PyJWT==2.8.0
python-dotenv==1.0.0
gevent==23.9.1
click==8.1.7

# --- ML & Audio Libraries from your project ---
torch
joblib
librosa
numpy
pydub
pyannote.audio
soundfile
matplotlib
demucs